Added convex hull rendering of connectors. Has not worked out quite as well (visually) as I had hoped, however I will use the generated convex hull to now create a minimal-bounding-box.
Added a level of part/pin-size intelligence based on the length of the part. So far rendering 0201~1206 caps/res, as well as diodes and some tant caps and inductors. Likely will need more refinement but it is a good visual start and the boards are now looking somewhat more realistic.