Changed the sleep/CPU-saving system.
Previously we queried SDL_GetWindowFlags() to see if the window was minimised or hidden
but there was some issues when the window was *shaded/shuttered* (ie, in Fluxbox) which
would cause the CPU to go quite high as OBV would be rendering the frames but not
displaying them, causing a tight run loop.

The new method sets a non-event countdown time, reset each time there is a new event.
Currently it is set to 3 frames before going to sleep (which is a 50ms snooze).